EventSystem + currency

+++ GEPProject/Source/GEPProject/Enemy/Enemy_Base.h
@@ -45,7 +45,7 @@ protected:
 	UParticleSystem* explosionParticles;
 
 	UPROPERTY(EditAnywhere)
-	float currencyToDrop;
+	int currencyToDrop;
 	
 	void Death();
 };
